+++ components/demo/src/memfault_demo_shell_commands.c
@@ -6,7 +6,7 @@
 //! @brief
 //! Command definitions for the minimal shell/console implementation.
 
-#include "memfault_demo_shell_commands.h"
+#include "memfault/demo/shell_commands.h"
 
 #include <stddef.h>
 
@@ -38,9 +38,7 @@ int memfault_demo_cli_cmd_crash(MEMFAULT_UNUSED int argc, MEMFAULT_UNUSED char *
 static const sMemfaultShellCommand s_memfault_shell_commands[] = {
   {"get_core", memfault_demo_cli_cmd_get_core, "Get coredump info"},
   {"clear_core", memfault_demo_cli_cmd_clear_core, "Clear an existing coredump"},
-  {"print_chunk", memfault_demo_cli_cmd_print_chunk, "Get next Memfault data chunk to send and print as a curl command"},
   {"crash", memfault_demo_cli_cmd_crash, "Trigger a crash"},
-  {"test_log", memfault_demo_cli_cmd_test_log, "Writes test logs to log buffer"},
   {"trigger_logs", memfault_demo_cli_cmd_trigger_logs, "Trigger capture of current log buffer contents"},
   {"drain_chunks",  memfault_demo_drain_chunk_data, "Flushes queued Memfault data. To upload data see https://mflt.io/posting-chunks-with-gdb"},
   {"trace", memfault_demo_cli_cmd_trace_event_capture, "Capture an example trace event"},