+++ components/include/memfault/default_config.h
@@ -21,9 +21,15 @@ extern "C" {
 // Core Components
 //
 
-//! Controls the
+//! Controls the use of the Gnu build ID feature.
 #ifndef MEMFAULT_USE_GNU_BUILD_ID
 #define MEMFAULT_USE_GNU_BUILD_ID 0
+#endif
+
+//! Allows users to dial in the correct amount of storage for their
+//! software version + build ID string.
+#ifndef MEMFAULT_UNIQUE_VERSION_MAX_LEN
+#define MEMFAULT_UNIQUE_VERSION_MAX_LEN 32
 #endif
 
 //! While it is recommended that MEMFAULT_SDK_ASSERT be left enabled, they can
@@ -100,6 +106,15 @@ extern "C" {
 //!   CFLAGS += -DMEMFAULT_EVENT_STORAGE_READ_BATCHING_ENABLED=1
 #ifndef MEMFAULT_EVENT_STORAGE_READ_BATCHING_ENABLED
 #define MEMFAULT_EVENT_STORAGE_READ_BATCHING_ENABLED 0
+#endif
+
+//! Enables support for the non-volatile event storage at compile time
+//! instead of dynamically at runtime
+//!
+//! Disabling this feature saves several hundred bytes of codespace and can be useful to enable for
+//! extremely constrained environments
+#ifndef MEMFAULT_EVENT_STORAGE_NV_SUPPORT_ENABLED
+#define MEMFAULT_EVENT_STORAGE_NV_SUPPORT_ENABLED 1
 #endif
 
 #if MEMFAULT_EVENT_STORAGE_READ_BATCHING_ENABLED != 0
@@ -279,6 +294,11 @@ extern "C" {
 #ifndef MEMFAULT_DEMO_CLI_USER_CHUNK_SIZE
 // Note: Arbitrary default size for CLI command. Can be as small as 9 bytes.
 #define MEMFAULT_DEMO_CLI_USER_CHUNK_SIZE 1024
+#endif
+
+//! The maximum length supported for a single CLI command
+#ifndef MEMFAULT_DEMO_SHELL_RX_BUFFER_SIZE
+#define MEMFAULT_DEMO_SHELL_RX_BUFFER_SIZE 64
 #endif
 
 //